test: registration unit tests
